requirements.txt
# ================================================================
# Astana Prices — зависимости Python
# ================================================================

# Web
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.3.0

# Redis / Celery
redis[hiredis]>=5.0.0
celery[redis]>=5.4.0
flower>=2.0.0

# Gemini AI
google-generativeai>=0.7.0

# Telegram
python-telegram-bot>=21.0.0

# Scraping
playwright>=1.44.0
parsel>=1.9.0

# Images (генерация карточек)
Pillow>=10.3.0

# Config
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Utils
pytz>=2024.1
structlog>=24.2.0
tenacity>=8.3.0

# Auth
python-jose[cryptography]>=3.3.0

# Monitoring
prometheus-fastapi-instrumentator>=7.0.0
